+++ engine/core/src/main/java/es/eucm/ead/engine/components/behaviors/events/RuntimeKey.java
@@ -125,7 +125,7 @@ public class RuntimeKey extends RuntimeBehavior {
 		this.shift = shift;
 	}
 
-	public boolean equals(Object o) {
+	public boolean compareEvents(Object o) {
 		if (o == null) {
 			return false;
 		}
